Rename PyTorch example
